# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText API by Loris
  description: "APIs for the WASAText project: just another messaging platform"
  version: 0.0.1
...
security:
  - bearerAuth: []
paths:
...
  /api/users/{userID}/conversations/users/{conversationID}/messages:
    parameters:
      - $ref: "#/components/parameters/userID"
      - $ref: "#/components/parameters/conversationID"

    get:
      tags:
        - Conversation
      summary: List messages in a conversation belong users.
      description: Returns messages for the specified conversationID belong users.
      operationId: getConversation
      responses:
        "200":
          description: All messages of conversation
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    description: List of messages.
                    type: array
                    minItems: 0
                    maxItems: 10000
                    items:
                      $ref: "#/components/schemas/Message"
        "404":
          description: Conversation or user not found
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags:
        - Message
      summary: Post a new message in the conversation belong users
      description: Enables the user to send a new message in a particular conversation belong users
      operationId: sendMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Message content (text or photo)
              required:
                - content
                - content_type
              properties:
                content:
                  description: Text or photo content.
                  oneOf:
                    - $ref: "#/components/schemas/Photo"
                    - $ref: "#/components/schemas/Text"
                content_type:
                  description: Type of the content message
                  type: string
                  enum:
                    - gif
                    - text
                    - photo
                  example: photo
                  pattern: "^(gif|text|photo)$"
      responses:
        "201":
          description: Message sent successfully
          content:
            application/json:
              schema:
                description: messageID of the new message
                type: object
                properties:
                  messageID:
                    $ref: "#/components/schemas/messageID"
        "400":
          description: Invalid message content.
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: Conversation not found.
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/users/{userID}/conversations/groups/{conversationID}/messages:
    parameters:
      - $ref: "#/components/parameters/userID"
      - $ref: "#/components/parameters/conversationID"

    get:
      tags:
        - Conversation
      summary: List messages in a conversation between users and groups.
      description: Returns messages for the specified conversationID between users and groups.
      operationId: getConversationGroups
      responses:
        "200":
          description: All messages of conversation
          content:
            application/json:
              schema:
                type: object
                description: Response containing the list of messages in the group conversation
                properties:
                  messages:
                    description: List of messages.
                    type: array
                    minItems: 0
                    maxItems: 10000
                    items:
                      $ref: "#/components/schemas/Message"
        "404":
          description: Conversation or user not found
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags:
        - Message
      summary: Post a new message in the conversation between users and groups
      description: Enables the user to send a new message in a particular conversation between users and groups
      operationId: sendMessageGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Message content for group conversation (text or photo)
              required:
                - content
                - content_type
              properties:
                content:
                  description: Text or photo content.
                  oneOf:
                    - $ref: "#/components/schemas/Photo"
                    - $ref: "#/components/schemas/Text"
                content_type:
                  description: Type of the content message
                  type: string
                  enum:
                    - gif
                    - text
                    - photo
                  example: photo
                  pattern: "^(gif|text|photo)$"
      responses:
        "201":
          description: Message sent successfully
          content:
            application/json:
              schema:
                description: messageID of the new message
                type: object
                properties:
                  messageID:
                    $ref: "#/components/schemas/messageID"
        "400":
          description: Invalid message content.
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: Conversation not found.
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/users/{userID}/conversations/users/{conversationID}/messages/{messageID}:
    parameters:
      - $ref: "#/components/parameters/userID" #userID of the user performing the action.
      - $ref: "#/components/parameters/conversationID"
      - $ref: "#/components/parameters/messageID"

    post:
      tags:
        - Message
      summary: Forward a message belong users
      description: Forwards an existing message to a target conversation belong users.
      operationId: forwardMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - source_type
              properties:
                source_type:
                  type: string
                  enum: ["users", "groups"]
                  description: Whether the original message comes from a user-user (users) or user-group (groups) conversation.
                  example: "users"
      responses:
        "201":
          description: Message forward successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  messageID:
                    $ref: "#/components/schemas/messageID"
                  replyToMessageID:
                    type: integer
                    description: The ID of the message this message is replying to.
                    example: 18
        "400":
          description:
            Invalid conversation or message id, or invalid target conversation
            id
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description:
            The original conversation or message could not be found, or
            the target conversation does not exist.
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      tags:
        - Message
      summary: Remove a message from a conversation belong users
      description: Enables the user to remove a specific message from a conversation belong users.
      operationId: deleteMessage
      responses:
        "200":
          description: Message deleted successfully
          content:
            application/json:
              schema:
                type: object
                description: Response containing the ID of the deleted message
                properties:
                  messageID:
                    $ref: "#/components/schemas/messageID"
        "400":
          description: Invalid message id or conversation id
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: Message not found
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: JWT bearer authentication scheme.
  schemas:
    Message:
      title: Message
      description: Represent a Message between users.
      type: object
      properties:
        content:
          description: Photo or text content.
          oneOf:
            - $ref: "#/components/schemas/Photo"
            - $ref: "#/components/schemas/Text"
        content_type:
          description: Type of the content message
          type: string
          enum:
            - gif
            - text
            - photo
          example: photo
          pattern: "^(gif|text|photo)$"
        timestamp:
          type: string
          format: date-time
          description: Message timestamp.
          example: "2025-10-25T17:40:11Z"
          pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}Z$'
        sender:
          $ref: "#/components/schemas/SimpleUser"
        status:
          type: string
          enum:
            - received
            - read
          description: Status of the sent message
          example: received
          pattern: "^(received|read)$"
        type:
          description: Type of the message
          type: string
          enum:
            - standard
            - forward
          example: standard
          pattern: "^(standard|forward)$"
        comments:
          type: array
          minItems: 0
          maxItems: 1000
          description: Array of comment for the message
          items:
            $ref: "#/components/schemas/Comment"
        messageID:
          $ref: "#/components/schemas/messageID"
        replyToMessageID:
          type: integer
          description: The ID of the message this message is replying to.
          example: 12
    Comment:
      title: Comment
      description: Represent a Comment for message.
      type: object
      properties:
        content:
          type: string
          description: emoji
          example: 👍
          minLength: 1
          maxLength: 2
          pattern: "^.*?$"
        sender:
          $ref: "#/components/schemas/SimpleUser"
        commentID:
          description: Unique comment identifier.
          type: integer
          example: 1
          readOnly: true
...
    SimpleUser:
      title: Simple User.
      description: Represents a User (without userID)
      type: object
      properties:
        userName:
          $ref: "#/components/schemas/userName"
...
    Photo:
      description: encoded image data
      type: string
      format: byte
      minLength: 0
      maxLength: 13981013
      example: aGVsbG8=
      pattern: "^[A-Za-z0-9+/]+={0,2}$"
...
    userName:
      description: The name of the user
      type: string
      example: loris2155519
      pattern: ^[a-z]+[0-9]*$
      minLength: 3
      maxLength: 15
    userID:
      description: Unique user identifier
      type: integer
      example: 1
    messageID:
      description: Unique message identifier.
      type: integer
      example: 1
...
    Text:
      type: string
      description: Text content of the message
      example: Prova messaggio
      minLength: 1
      maxLength: 4095
      pattern: ^.*?$
  parameters:
    userID:
      schema:
        $ref: "#/components/schemas/userID"
      name: userID
      in: path
      required: true
      description: User identifier
    conversationID:
      schema:
        type: integer
        example: 1
        readOnly: true
        description: Unique identifier of the conversation.
      name: conversationID
      in: path
      required: true
    messageID:
      schema:
        type: integer
        example: 1
        readOnly: true
        description: Unique identifier for the message.
      name: messageID
      in: path
      required: true
      description: Message id
...
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
    InternalServerError:
      description: Internal server error